Moved the sorter code into separate trait
