+++ src/config.rs
@@ -3,6 +3,8 @@ use std::marker::PhantomData;
 
 use derivative::*;
 
+use crate::{dag::ResolverSorter, field::SmallField};
+
 // pub const DEBUG_SATISFIABLE: bool = true;
 pub const DEBUG_SATISFIABLE: bool = false;
 
@@ -22,6 +24,7 @@ pub trait CSSetupConfig: 'static + Send + Sync + Clone + Copy + std::fmt::Debug 
 
 pub trait CSResolverConfig: 'static + Send + Sync + Clone + Copy + std::fmt::Debug {
     type DebugConfig: CSDebugConfig;
+    type Sorter<F: SmallField>: ResolverSorter<F>;
 }
 
 pub trait CSConfig: 'static + Send + Sync + Clone + Copy + std::fmt::Debug {
@@ -85,6 +88,8 @@ pub struct Resolver<Dbg: CSDebugConfig>(PhantomData<Dbg>);
 
 impl<Dbg: CSDebugConfig> CSResolverConfig for Resolver<Dbg> {
     type DebugConfig = Dbg;
+
+    type Sorter<F: SmallField> = crate::dag::sorter_runtime::RuntimeResolverSorter<F, Self>;
 }
 
 #[derive(Derivative)]
