+++ src/dag/mod.rs
@@ -1,7 +1,11 @@
+use std::cell::UnsafeCell;
 use std::hint::spin_loop;
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
+use std::sync::{Arc, Mutex};
+use std::thread::JoinHandle;
 
+use crate::config::CSResolverConfig;
+use crate::cs::traits::cs::DstBuffer;
 use crate::field::SmallField;
 
 mod awaiters;
@@ -10,6 +14,7 @@ mod registrar;
 pub(crate) mod resolution_window;
 pub(crate) mod resolver;
 mod resolver_box;
+pub(crate) mod sorter_runtime;
 
 pub trait TrivialWitnessCastable<F: SmallField, const N: usize>:
     'static + Clone + std::fmt::Debug + Send + Sync
@@ -75,6 +80,10 @@ impl<F: SmallField, const N: usize, S: WitnessSource<F>> CSWitnessValues<F, N, S
 
 use crate::cs::Place;
 
+use self::guide::{GuideOrder, OrderInfo};
+use self::resolver::{CircuitResolverOpts, ResolverIx, ResolverCommonData};
+use self::resolver_box::ResolverBox;
+
 // we use Arc and interior mutability, so we want Send + Sync just in case
 pub trait WitnessSource<F: SmallField>: 'static + Send + Sync {
     const PRODUCES_VALUES: bool;
@@ -91,4 +100,30 @@ pub trait WitnessSourceAwaitable<F: SmallField>: WitnessSource<F> {
 
 pub trait Awaiter<'a> {
     fn wait(&self);
+}
+
+pub trait ResolverSorter<F: SmallField>: Sized {
+
+    fn new(opts: CircuitResolverOpts, debug_track: &Vec<Place>) -> (Self, Arc<ResolverCommonData<F>>);
+    fn set_value(&mut self, key: Place, value: F);
+    fn add_resolution<Fn>(&mut self, inputs: &[Place], outputs: &[Place], f: Fn)
+    where
+        Fn: FnOnce(&[F], &mut DstBuffer<'_, '_, F>) + Send + Sync;
+
+    unsafe fn internalize(&mut self, resolver_ix: ResolverIx, inputs: &[Place], outputs: &[Place]);
+    fn internalize_one(
+        &mut self,
+        resolver_ix: ResolverIx,
+        inputs: &[Place],
+        outputs: &[Place],
+    ) -> Vec<ResolverIx>;
+
+    fn write_order<'a, T: GuideOrder<'a, ResolverIx>>(
+        tgt: &Mutex<Vec<OrderInfo<ResolverIx>>>,
+        resolvers: &UnsafeCell<ResolverBox<F>>,
+        order: &T,
+    );
+
+    fn flush(&mut self);
+    fn final_flush(&mut self);
 }
