+++ src/dag/resolver.rs
@@ -2,6 +2,7 @@
 #![allow(clippy::overly_complex_bool_expr)]
 #![allow(clippy::nonminimal_bool)]
 
+use super::ResolverSorter;
 use crate::log;
 use std::any::Any;
 use std::cell::{Cell, UnsafeCell};
@@ -38,7 +39,7 @@ pub struct CircuitResolverOpts {
 }
 
 /// Shared between the resolver, awaiters and the resolution window.
-pub(crate) struct ResolverCommonData<V> {
+pub struct ResolverCommonData<V> {
     // The following two are meant to have an asynchronized access. The access
     // correctness is provided by the `exec_order` mutex. Once an iter is
     // added to the `exec_order`, it's guaranteed that the corresponding
@@ -52,7 +53,6 @@ pub(crate) struct ResolverCommonData<V> {
     /// amount of resolvers.
     pub exec_order: Mutex<Vec<OrderInfo<ResolverIx>>>,
     pub awaiters_broker: AwaitersBroker,
-    pub comms: ResolverComms,
 }
 
 /// Used to send notifications and data between the resolver, resolution window
@@ -97,15 +97,20 @@ impl Stats {
 ///    Unsafe to work with.
 
 pub struct CircuitResolver<V: SmallField, Cfg: CSResolverConfig> {
-    registrar: Registrar,
+    // registrar: Registrar,
+
+    sorter: Cfg::Sorter<V>,
 
     pub(crate) common: Arc<ResolverCommonData<V>>,
     pub(crate) options: CircuitResolverOpts,
-    pub(crate) guide: BufferGuide<ResolverIx, Cfg>,
+    
+    comms: Arc<ResolverComms>,
+    // pub(crate) guide: BufferGuide<ResolverIx, Cfg>,
 
     resolution_window_handle: Option<JoinHandle<()>>,
 
     stats: Stats,
+    call_count: u32,
     debug_track: Vec<Place>,
 }
 
@@ -113,24 +118,9 @@ unsafe impl<V: SmallField, Cfg: CSResolverConfig> Send for CircuitResolver<V, Cf
 unsafe impl<V: SmallField, Cfg: CSResolverConfig> Sync for CircuitResolver<V, Cfg> where V: Send {}
 
 // TODO: try to eliminate this constraint to something more general, preferably defatult.
-impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
+impl<V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
     pub fn new(opts: CircuitResolverOpts) -> Self {
-        fn new_values<V>(size: usize, default: fn() -> V) -> Box<[V]> {
-            // TODO: ensure mem-page multiple capacity.
-            let mut values = Vec::with_capacity(size);
-            // TODO: If this isn't reused extend should be switched to unsafe resize
-            values.resize_with(size, default);
-            values.into_boxed_slice()
-        }
-
-        let values = Values {
-            variables: new_values(opts.max_variables, || {
-                UnsafeCell::new((V::from_u64_unchecked(0), Metadata::default()))
-            }),
-            max_tracked: -1,
-        };
-
-        let exec_order = Vec::with_capacity(opts.max_variables);
+        let threads = 3;
 
         let debug_track = vec![];
 
@@ -138,28 +128,25 @@ impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
             log!("Contains tracked keys {:?} ", debug_track);
         }
 
-        let common = ResolverCommonData {
-            resolvers: UnsafeCell::new(ResolverBox::new()),
-            values: UnsafeCell::new(values),
-            exec_order: Mutex::new(exec_order),
-            awaiters_broker: AwaitersBroker::new(),
-            comms: ResolverComms {
-                registration_complete: AtomicBool::new(false),
-                rw_panicked: AtomicBool::new(false),
-                rw_panic: Cell::new(None),
-            },
-        }
-        .to(Arc::new);
+        let (sorter, common) = Cfg::Sorter::new(opts, &debug_track);
 
-        let threads = 3;
+        let comms = ResolverComms {
+            registration_complete: AtomicBool::new(false),
+            rw_panicked: AtomicBool::new(false),
+            rw_panic: Cell::new(None),
+        }.to(Arc::new);
 
         Self {
             options: opts,
-            guide: BufferGuide::new(opts.desired_parallelism),
-            registrar: Registrar::new(),
+            call_count: 0,
+            sorter,
+            // guide: BufferGuide::new(opts.desired_parallelism),
+            // registrar: Registrar::new(),
+            comms: comms.clone(),
 
             resolution_window_handle: ResolutionWindow::<V>::run(
-                Arc::clone(&common),
+                comms.clone(),
+                common.clone(),
                 &debug_track,
                 threads,
             )
@@ -172,294 +159,176 @@ impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
     }
 
     pub fn set_value(&mut self, key: Place, value: V) {
-        if (cfg!(cr_paranoia_mode) || PARANOIA) && self.debug_track.contains(&key) && false {
-            log!("CR: setting {:?} -> {:?}", key, value);
-        }
-
-        match key.get_type() {
-            VariableType::CopyableVariable => self.stats.values_added += 1,
-            VariableType::Witness => self.stats.witnesses_added += 1,
-        }
-
-        // Safety: Dereferencing as &mut in mutable context. This thread doesn't hold any
-        // references to `self.resolvers`. Other thread may hold shared references, but
-        // are guaranteed to not access the same underlying data.
-        let values = unsafe { self.common.values.u_deref_mut() };
-
-        values.set_value(key, value);
-
-        // Safety: using as shared, assuming no &mut references to
-        // `self.resolvers` (Only this thread requires mut, and we're not
-        // currently doing that).
-
-        let delayed_resolvers =
-            if values.max_tracked >= 0 {
-                self.registrar.advance(values.max_tracked.to(|x| {
-                    Place::from_variable(Variable::from_variable_index(x.try_into().unwrap()))
-                }))
-            } else {
-                vec![]
-            };
-
-        unsafe {
-            // Safety: Dereferencing as shared, not accessing `resolve_fn`.
-            let rb = self.common.resolvers.u_deref();
-
-            delayed_resolvers
-                .into_iter()
-                .map(|x| (x, rb.get(x).inputs(), rb.get(x).outputs()))
-                // Safety: No &mut references to `self.common.resolvers`.
-                .for_each(|(r, i, o)| self.internalize(r, i, o));
-        }
+        self.sorter.set_value(key, value)
     }
 
     pub fn add_resolution<F>(&mut self, inputs: &[Place], outputs: &[Place], f: F)
     where
-        F: FnOnce(&[V], &mut DstBuffer<'_, '_, V>) + Send + Sync + 'cr,
+        F: FnOnce(&[V], &mut DstBuffer<'_, '_, V>) + Send + Sync,
     {
-        debug_assert!(inputs
-            .iter()
-            .all(|x| x.0 < self.options.max_variables as u64));
-
-        // Safety: This thread is the only one to use `push` on the resolvers
-        // and is the only thread to do so. `push` is the only mutable function
-        // on that struct.
-        let resolver_ix = unsafe {
-            self.common
-                .resolvers
-                .u_deref_mut()
-                .push(inputs, outputs, f, invocation_binder::<F, V>)
-        };
-
-        if PARANOIA && resolver_ix.0 == 0 {
-            println!(
-                "CR: Resolvers push returned ix 0, on resolution {}",
-                self.stats.registrations_added
-            );
-        }
-
-        let mut hit = false;
-
-        if (cfg!(cr_paranoia_mode) || PARANOIA) && true {
-            if let Some(x) = self.debug_track.iter().find(|x| inputs.contains(x)) {
-                log!("CR: added resolution with tracked input {:?}", x);
-
-                hit = true;
-            }
-
-            if let Some(x) = self.debug_track.iter().find(|x| outputs.contains(x)) {
-                log!("CR: added resolution with tracked output {:?}", x);
-
-                hit = true;
-            }
-
-            if hit {
-                log!("   {:?}", resolver_ix);
-                log!(
-                    "   Ins:\n   - {}{}\n   Outs:\n   - {}{}",
-                    inputs
-                        .iter()
-                        .take(10)
-                        .map(|x| format!("{:?}", x))
-                        .collect_vec()
-                        .join("\n   - "),
-                    if inputs.len() > 10 {
-                        format!("\n   - ... {} total", inputs.len())
-                    } else {
-                        "".to_owned()
-                    },
-                    outputs
-                        .iter()
-                        .take(10)
-                        .map(|x| format!("{:?}", x))
-                        .collect_vec()
-                        .join("\n   - "),
-                    if outputs.len() > 10 {
-                        format!("\n   - ... {} total", outputs.len())
-                    } else {
-                        "".to_owned()
-                    },
-                );
-            }
-        }
-
-        let registrar_answer = self.registrar.accept(inputs, resolver_ix);
-
-        if hit {
-            match registrar_answer {
-                Err(x) => log!("   Registration delayed due to {:?}", x),
-                Ok(_) => log!("   Registration accepted."),
-            }
-        }
-
-        if let Ok(resolver_ix) = registrar_answer {
-            // Safety: `self.resolvers` is dropped as a temp value a few lines above.
-            unsafe { self.internalize(resolver_ix, inputs, outputs) };
-        }
-
-        self.stats.registrations_added += 1;
+        self.sorter.add_resolution(inputs, outputs, f)
     }
 
     /// Safety: `self.common.resolvers` must not have a &mut reference to it.
-    unsafe fn internalize(&mut self, resolver_ix: ResolverIx, inputs: &[Place], outputs: &[Place]) {
-        let mut resolvers = vec![(resolver_ix, inputs, outputs)];
-
-        if PARANOIA && resolver_ix == ResolverIx::new_resolver(0) {
-            println!("CR: Internalize called with resolver_ix 0");
-        }
-
-        // Safety: using as shared, assuming no &mut references to
-        // `self.resolvers` that access the same underlying data. This is
-        // guaranteed by the fact that inputs and outputs for any given resolver
-        // is written only once by this thread, and safety requires that no &mut
-        // references to `self.resolvers` exist.
-        let rb = self.common.resolvers.u_deref();
-
-        while resolvers.len() > 0 {
-            let (resolver_ix, inputs, outputs) = resolvers.pop().unwrap();
-
-            let new_resolvers = self.internalize_one(resolver_ix, inputs, outputs);
-
-            if PARANOIA {
-                if new_resolvers.iter().any(|x| x.0 == 0) {
-                    println!("CR: internalize_one returned resolver with ix 0");
-                }
-            }
-
-            self.registrar.stats.secondary_resolutions += new_resolvers.len();
-
-            // Safety: calling to immutable functions (`get`, `inputs`, `outputs`).
-            // The resolver is not yet pushed to the resolution window.
-            new_resolvers
-                .into_iter()
-                .map(|x| (x, rb.get(x).inputs(), rb.get(x).outputs()))
-                .to(|x| resolvers.extend(x));
-        }
-    }
-
-    fn internalize_one(
-        &mut self,
-        resolver_ix: ResolverIx,
-        inputs: &[Place],
-        outputs: &[Place],
-    ) -> Vec<ResolverIx> {
-        if cfg!(cr_paranoia_mode) {
-            if let Some(x) = self.debug_track.iter().find(|x| inputs.contains(x)) {
-                log!("CR: internalized resolution with tracked input {:?}", x);
-            }
-
-            if let Some(x) = self.debug_track.iter().find(|x| outputs.contains(x)) {
-                log!("CR: internalized resolution with tracked output {:?}", x);
-            }
-        }
-
-        // Safety: The values created by this function are not yet tracked, thus
-        // are not referenced by anyone. All dependencies have already been
-        // written.
-        // It is safe to borrow as mut because the only mutable function that is
-        // being called on it is `track_values` and thus no one can have a
-        // reference to that location.
-        // The rest of the calls are immutable.
-        let values = unsafe { self.common.values.u_deref_mut() };
-
-        let deps = inputs.iter().map(|x| &values.get_item_ref(*x).1);
-
-        if cfg!(cr_paranoia_mode) {
-            debug_assert!(
-                deps.clone().all(|x| { x.is_tracked() }),
-                "Attempting to internalize a resolution with an untracked input. All inputs must be tracked."
-            );
-        }
-
-        if PARANOIA && resolver_ix == ResolverIx::new_resolver(0) && false {
-            self.guide.tracing = true;
-        }
-
-        if PARANOIA && resolver_ix == ResolverIx::new_resolver(0) {
-            println!("CR: resolver_ix {} pushed to guide.", resolver_ix.0);
-        }
-
-        let (guide_loc, order) = self
-            .guide
-            .push(resolver_ix, deps.map(|x| x.guide_loc).reduce(std::cmp::max));
-
-        Self::write_order(&self.common.exec_order, &self.common.resolvers, &order);
-
-        values.track_values(outputs, guide_loc);
-
-        // This values starts from -1, which is illegal.
-        if values.max_tracked >= 0 {
-            let delayed_resolvers = self.registrar.advance(values.max_tracked.to(|x| {
-                Place::from_variable(Variable::from_variable_index(x.try_into().unwrap()))
-            }));
-
-            delayed_resolvers
-        } else {
-            Vec::new()
-        }
-    }
-
-    fn write_order<'a, T: GuideOrder<'a, ResolverIx>>(
-        tgt: &Mutex<Vec<OrderInfo<ResolverIx>>>,
-        resolvers: &UnsafeCell<ResolverBox<V>>,
-        order: &T,
-    ) {
-        if order.size() > 0 {
-            let mut tgt = tgt.lock().unwrap();
-            let len = tgt.len();
-            tgt.resize(
-                len + order.size(),
-                OrderInfo::new(ResolverIx::default(), GuideMetadata::new(0)),
-            );
-
-            order.write(&mut tgt[..]);
-
-            if PARANOIA {
-                for i in len..len + order.size() {
-                    if tgt[i].value == ResolverIx(0) {
-                        log!(
-                            "CR: resolver {} added to order at ix {}, during write {}..{}.",
-                            tgt[i].value.0,
-                            i,
-                            len,
-                            len + order.size()
-                        );
-                    }
-                }
-            }
-
-            if cfg!(cr_paranoia_mode) {
-                // This ugly block checks that the calculated parallelism is
-                // correct. It's a bit slower than O(n^2). Also note, that it
-                // checks only the last 1050 items, so it's not a full check,
-                // 'twas when the desired parallelism was set to 1024, but it's
-                // not anymore.
-                unsafe {
-                    for r_ix in std::cmp::max(0, len as i32 - 1050) as usize..tgt.len() {
-                        let r = resolvers.u_deref().get(tgt[r_ix].value);
-
-                        for derivative in
-                            r_ix..std::cmp::min(r_ix + tgt[r_ix].metadata.parallelism(), tgt.len())
-                        {
-                            let r_d = resolvers.u_deref().get(tgt[derivative].value);
-
-                            assert!(r.outputs().iter().all(|x| r_d.inputs().contains(x) == false),
-                                "Parallelism violation at ix {}, val {:#?}, derivative ix {} , val {:#?}: {:#?}",
-                                r_ix,
-                                tgt[r_ix],
-                                derivative,
-                                tgt[derivative],
-                                (std::cmp::max(0, len as i32 - 1050) as usize..tgt.len())
-                                    .map(|x| (x, resolvers.u_deref().get(tgt[x].value)))
-                                    .map(|(i, r)| (i, tgt[i], r.inputs().to_vec(), r.outputs().to_vec()))
-                                    .collect_vec()
-                            );
-                        }
-                    }
-                }
-            }
-        }
-    }
+    // unsafe fn internalize(&mut self, resolver_ix: ResolverIx, inputs: &[Place], outputs: &[Place]) {
+    //     let mut resolvers = vec![(resolver_ix, inputs, outputs)];
+    //
+    //     if PARANOIA && resolver_ix == ResolverIx::new_resolver(0) {
+    //         println!("CR: Internalize called with resolver_ix 0");
+    //     }
+    //
+    //     // Safety: using as shared, assuming no &mut references to
+    //     // `self.resolvers` that access the same underlying data. This is
+    //     // guaranteed by the fact that inputs and outputs for any given resolver
+    //     // is written only once by this thread, and safety requires that no &mut
+    //     // references to `self.resolvers` exist.
+    //     let rb = self.common.resolvers.u_deref();
+    //
+    //     while resolvers.len() > 0 {
+    //         let (resolver_ix, inputs, outputs) = resolvers.pop().unwrap();
+    //
+    //         let new_resolvers = self.internalize_one(resolver_ix, inputs, outputs);
+    //
+    //         if PARANOIA {
+    //             if new_resolvers.iter().any(|x| x.0 == 0) {
+    //                 println!("CR: internalize_one returned resolver with ix 0");
+    //             }
+    //         }
+    //
+    //         self.registrar.stats.secondary_resolutions += new_resolvers.len();
+    //
+    //         // Safety: calling to immutable functions (`get`, `inputs`, `outputs`).
+    //         // The resolver is not yet pushed to the resolution window.
+    //         new_resolvers
+    //             .into_iter()
+    //             .map(|x| (x, rb.get(x).inputs(), rb.get(x).outputs()))
+    //             .to(|x| resolvers.extend(x));
+    //     }
+    // }
+    //
+    // fn internalize_one(
+    //     &mut self,
+    //     resolver_ix: ResolverIx,
+    //     inputs: &[Place],
+    //     outputs: &[Place],
+    // ) -> Vec<ResolverIx> {
+    //     if cfg!(cr_paranoia_mode) {
+    //         if let Some(x) = self.debug_track.iter().find(|x| inputs.contains(x)) {
+    //             log!("CR: internalized resolution with tracked input {:?}", x);
+    //         }
+    //
+    //         if let Some(x) = self.debug_track.iter().find(|x| outputs.contains(x)) {
+    //             log!("CR: internalized resolution with tracked output {:?}", x);
+    //         }
+    //     }
+    //
+    //     // Safety: The values created by this function are not yet tracked, thus
+    //     // are not referenced by anyone. All dependencies have already been
+    //     // written.
+    //     // It is safe to borrow as mut because the only mutable function that is
+    //     // being called on it is `track_values` and thus no one can have a
+    //     // reference to that location.
+    //     // The rest of the calls are immutable.
+    //     let values = unsafe { self.common.values.u_deref_mut() };
+    //
+    //     let deps = inputs.iter().map(|x| &values.get_item_ref(*x).1);
+    //
+    //     if cfg!(cr_paranoia_mode) {
+    //         debug_assert!(
+    //             deps.clone().all(|x| { x.is_tracked() }),
+    //             "Attempting to internalize a resolution with an untracked input. All inputs must be tracked."
+    //         );
+    //     }
+    //
+    //     if PARANOIA && resolver_ix == ResolverIx::new_resolver(0) && false {
+    //         self.guide.tracing = true;
+    //     }
+    //
+    //     if PARANOIA && resolver_ix == ResolverIx::new_resolver(0) {
+    //         println!("CR: resolver_ix {} pushed to guide.", resolver_ix.0);
+    //     }
+    //
+    //     let (guide_loc, order) = self
+    //         .guide
+    //         .push(resolver_ix, deps.map(|x| x.guide_loc).reduce(std::cmp::max));
+    //
+    //     Self::write_order(&self.common.exec_order, &self.common.resolvers, &order);
+    //
+    //     values.track_values(outputs, guide_loc);
+    //
+    //     // This values starts from -1, which is illegal.
+    //     if values.max_tracked >= 0 {
+    //         let delayed_resolvers = self.registrar.advance(values.max_tracked.to(|x| {
+    //             Place::from_variable(Variable::from_variable_index(x.try_into().unwrap()))
+    //         }));
+    //
+    //         delayed_resolvers
+    //     } else {
+    //         Vec::new()
+    //     }
+    // }
+    //
+    // fn write_order<'a, T: GuideOrder<'a, ResolverIx>>(
+    //     tgt: &Mutex<Vec<OrderInfo<ResolverIx>>>,
+    //     resolvers: &UnsafeCell<ResolverBox<V>>,
+    //     order: &T,
+    // ) {
+    //     if order.size() > 0 {
+    //         let mut tgt = tgt.lock().unwrap();
+    //         let len = tgt.len();
+    //         tgt.resize(
+    //             len + order.size(),
+    //             OrderInfo::new(ResolverIx::default(), GuideMetadata::new(0)),
+    //         );
+    //
+    //         order.write(&mut tgt[..]);
+    //
+    //         if PARANOIA {
+    //             for i in len..len + order.size() {
+    //                 if tgt[i].value == ResolverIx(0) {
+    //                     log!(
+    //                         "CR: resolver {} added to order at ix {}, during write {}..{}.",
+    //                         tgt[i].value.0,
+    //                         i,
+    //                         len,
+    //                         len + order.size()
+    //                     );
+    //                 }
+    //             }
+    //         }
+    //
+    //         if cfg!(cr_paranoia_mode) {
+    //             // This ugly block checks that the calculated parallelism is
+    //             // correct. It's a bit slower than O(n^2). Also note, that it
+    //             // checks only the last 1050 items, so it's not a full check,
+    //             // 'twas when the desired parallelism was set to 1024, but it's
+    //             // not anymore.
+    //             unsafe {
+    //                 for r_ix in std::cmp::max(0, len as i32 - 1050) as usize..tgt.len() {
+    //                     let r = resolvers.u_deref().get(tgt[r_ix].value);
+    //
+    //                     for derivative in
+    //                         r_ix..std::cmp::min(r_ix + tgt[r_ix].metadata.parallelism(), tgt.len())
+    //                     {
+    //                         let r_d = resolvers.u_deref().get(tgt[derivative].value);
+    //
+    //                         assert!(r.outputs().iter().all(|x| r_d.inputs().contains(x) == false),
+    //                             "Parallelism violation at ix {}, val {:#?}, derivative ix {} , val {:#?}: {:#?}",
+    //                             r_ix,
+    //                             tgt[r_ix],
+    //                             derivative,
+    //                             tgt[derivative],
+    //                             (std::cmp::max(0, len as i32 - 1050) as usize..tgt.len())
+    //                                 .map(|x| (x, resolvers.u_deref().get(tgt[x].value)))
+    //                                 .map(|(i, r)| (i, tgt[i], r.inputs().to_vec(), r.outputs().to_vec()))
+    //                                 .collect_vec()
+    //                         );
+    //                     }
+    //                 }
+    //             }
+    //         }
+    //     }
+    // }
 
     pub fn wait_till_resolved(&mut self) {
         self.wait_till_resolved_impl(true);
@@ -467,7 +336,6 @@ impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
 
     pub fn wait_till_resolved_impl(&mut self, report: bool) {
         if self
-            .common
             .comms
             .registration_complete
             .load(std::sync::atomic::Ordering::Relaxed)
@@ -475,25 +343,11 @@ impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
             return;
         }
 
-        assert!(self.registrar.is_empty());
-        // assert!(self.registrar.is_empty(), "Registrar is not empty: has {:?}", self.registrar.peek_vars());
-
-        let order = self.guide.flush();
-
-        Self::write_order(&self.common.exec_order, &self.common.resolvers, &order);
-
-        drop(order);
-
-        if cfg!(cr_paranoia_mode) || PARANOIA {
-            log!(
-                "CR: Final order written. Order len {}",
-                self.common.exec_order.lock().unwrap().len()
-            );
-        }
+        self.sorter.flush();
 
         self.stats.registration_time = self.stats.started_at.elapsed();
 
-        self.common
+        self
             .comms
             .registration_complete
             .store(true, std::sync::atomic::Ordering::Relaxed);
@@ -508,12 +362,11 @@ impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
 
         // Propage panic from the resolution window handler.
         if self
-            .common
             .comms
             .rw_panicked
             .load(std::sync::atomic::Ordering::Relaxed)
         {
-            if let Some(e) = self.common.comms.rw_panic.take() {
+            if let Some(e) = self.comms.rw_panic.take() {
                 resume_unwind(e);
             } else {
                 log!("Resolution window panicked, but no panic payload stored.");
@@ -532,10 +385,6 @@ impl<'cr, V: SmallField, Cfg: CSResolverConfig> CircuitResolver<V, Cfg> {
         }
 
         if cfg!(cr_paranoia_mode) || PARANOIA {
-            self.guide.stats.finalize();
-
-            log!("CR {:?}", self.guide.stats);
-            log!("CRR stats {:#?}", self.registrar.stats);
             log!("CR {:?}", unsafe {
                 self.common.awaiters_broker.stats.u_deref()
             });
@@ -604,15 +453,11 @@ impl<V: SmallField, Cfg: CSResolverConfig> WitnessSourceAwaitable<V> for Circuit
 
         let r = awaiters::AwaitersBroker::register(
             &self.common.awaiters_broker,
-            &self.common.comms,
+            &self.comms,
             md,
         );
 
-        let order = self.guide.flush();
-
-        Self::write_order(&self.common.exec_order, &self.common.resolvers, &order);
-
-        drop(order);
+        self.sorter.flush();
 
         r
     }
@@ -636,9 +481,9 @@ impl<V: SmallField, Cfg: CSResolverConfig> Drop for CircuitResolver<V, Cfg> {
 // region: ResolverIx
 
 #[derive(Copy, Clone, Default, PartialEq, Eq, PartialOrd, Ord, Debug)]
-pub(crate) struct ResolverIx(pub u32);
+pub struct ResolverIx(pub u32);
 
-pub(crate) enum ResolverIxType {
+pub enum ResolverIxType {
     Jump,
     Resolver,
 }
@@ -720,7 +565,7 @@ impl<V> Values<V> {
 
     /// Marks values as tracked and stores the resolution order that those values
     /// are resolved in.
-    fn track_values(&mut self, keys: &[Place], loc: GuideLoc) {
+    pub(crate) fn track_values(&mut self, keys: &[Place], loc: GuideLoc) {
         for key in keys {
             let nmd = Metadata::new(loc);
 
@@ -739,7 +584,7 @@ impl<V> Values<V> {
         self.advance_track();
     }
 
-    fn set_value(&mut self, key: Place, value: V) {
+    pub(crate) fn set_value(&mut self, key: Place, value: V) {
         // Safety: we're setting the value, so we're sure that the item at key is not used.
         // If the item was already set, we panic in the next line.
         let (v, md) = unsafe { self.get_item_ref_mut(key) };
